Fix `xieite::memoize()`

+++ include/xieite/math/hash_combine.hpp
@@ -9,7 +9,7 @@
 namespace xieite {
 	template<std::integral T = std::size_t, std::convertible_to<T>... Ts>
 	[[nodiscard]] constexpr T hash_combine(T first, Ts... rest) noexcept {
-		(..., (first = std::rotl(first, xieite::bit_size<T> / 3) ^ xieite::hash_dist(static_cast<T>(rest))));
+		(..., (first = std::rotl(first, static_cast<int>(xieite::bit_size<T> / 3)) ^ xieite::hash_dist(static_cast<T>(rest))));
 		return first;
 	}
 }
